+++ CHANGELOG.md
@@ -27,6 +27,8 @@
 
 ### Changed
 
+- Refactored `_renderFrame` usage to instead implement the new `_getFrameButtons`. (#1863)
+
 ### Deprecated
 
 ### Removed